backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.9.0
python-dotenv==1.0.0
httpx==0.27.2
groq==0.11.0
fpdf2==2.7.6
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0
requests==2.31.0
reportlab>=4.4.3
slowapi==0.1.9
redis==5.0.1
python-docx==1.1.0
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml>=6.0.0
Pillow==10.1.0
matplotlib==3.8.2
numpy>=1.26.0
aiofiles==23.2.1
pydantic-settings==2.1.0
gitpython==3.1.40
PyGithub==2.1.1
chardet==5.2.0
# Universal document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.4
faker==20.1.0